Fixing missing protocol in the address
With the changes in fix-notls, the protocol becomes part of the node-ID, because it is needed
to make the correct connection and verify or not the certificate.

This change makes sure that the full name is used wherever necessary, and makes the tests
pass again.

There might still be things lingering which are not happy with the new node-ID.

Locally the tests with 'run_local.sh' pass.

Now all addresses need to have a port, even https://hostname.